requirements.txt
flask
transformers
torch
nats-py
numpy
aiohttp
pandas
tiktoken
protobuf
blobfile

